# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nestling API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/children/{id}/summary": {
      "get": {
        "tags": [
          "usage"
        ],
        "operationId": "summary",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Child id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "tz",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One local day, summarised",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such child in this family"
          },
          "422": {
            "description": "Unknown timezone"
          }
        }
      }
    },
    "/v1/children/{id}/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "operationId": "usage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Child id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tz",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage series for the requested local dates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such child in this family"
          },
          "422": {
            "description": "Unknown timezone"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceStatus": {
        "type": "object",
        "required": [
          "id",
          "label",
          "stale"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "last_seen_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "stale": {
            "type": "boolean"
          }
        }
      },
      "DeviceTotal": {
        "type": "object",
        "required": [
          "start",
          "screen_on_ms",
          "unlock_count"
        ],
        "properties": {
          "screen_on_ms": {
            "type": "integer",
            "format": "int64"
          },
          "start": {
            "type": "string"
          },
          "unlock_count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Point": {
        "type": "object",
        "required": [
          "start",
          "foreground_ms",
          "launch_count"
        ],
        "properties": {
          "foreground_ms": {
            "type": "integer",
            "format": "int64"
          },
          "launch_count": {
            "type": "integer",
            "format": "int32"
          },
          "start": {
            "type": "string",
            "description": "RFC3339 instant for hourly buckets, `YYYY-MM-DD` for daily ones."
          }
        }
      },
...
      "Series": {
        "type": "object",
        "required": [
          "package",
          "label",
          "points"
        ],
        "properties": {
          "label": {
            "type": "string"
          },
          "package": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Point"
            }
          }
        }
      },
      "SummaryResponse": {
        "type": "object",
        "required": [
          "child_id",
          "date",
          "tz",
          "total_ms",
          "unlock_count",
          "top_apps"
        ],
        "properties": {
          "child_id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "first_activity": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_activity": {
            "type": [
              "string",
              "null"
            ]
          },
          "top_apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TopApp"
            }
          },
          "total_ms": {
            "type": "integer",
            "format": "int64"
          },
          "tz": {
            "type": "string"
          },
          "unlock_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TopApp": {
        "type": "object",
        "required": [
          "package",
          "label",
          "foreground_ms"
        ],
        "properties": {
          "foreground_ms": {
            "type": "integer",
            "format": "int64"
          },
          "label": {
            "type": "string"
          },
          "package": {
            "type": "string"
          }
        }
      },
      "UsageResponse": {
        "type": "object",
        "required": [
          "child_id",
          "from",
          "to",
          "bucket",
          "tz",
          "devices",
          "series",
          "device_totals"
        ],
        "properties": {
          "bucket": {
            "type": "string"
          },
          "child_id": {
            "type": "string",
            "format": "uuid"
          },
          "device_totals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceTotal"
            }
          },
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceStatus"
            }
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Series"
            }
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "tz": {
            "type": "string"
          }
        }
      }
    }
  }
}